# frontend/openapi.json
{"openapi":"3.1.0","info":{"title":"Schoen Macht Geld API","description":"Backend for the stock exchange party game","version":"0.1.0"},"servers":[{"url":"/api"}],"paths":{"/stocks/":{"get":{"tags":["stocks"],"summary":"List Stocks","description":"Get all stocks.\n\nArgs:\n    order: Ordering option (default, random, rank, rank_desc, created_at,\n    created_at_desc, change_rank, change_rank_desc)\n    limit: Maximum number of stocks to return","operationId":"list_stocks_stocks__get","parameters":[{"name":"order","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/StockOrder"},{"type":"null"}],"title":"Order"}},{"name":"limit","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockResponse"},"title":"Response List Stocks Stocks  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["stocks"],"summary":"Create Stock","description":"Create a new stock.","operationId":"create_stock_stocks__post","requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_create_stock_stocks__post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stocks/{ticker}":{"get":{"tags":["stocks"],"summary":"Get Stock","description":"Get a single stock by ticker.","operationId":"get_stock_stocks__ticker__get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stocks/{ticker}/image":{"post":{"tags":["stocks"],"summary":"Upload Stock Image","description":"Upload and store stock image locally.","operationId":"upload_stock_image_stocks__ticker__image_post","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_stock_image_stocks__ticker__image_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stocks/{ticker}/price":{"post":{"tags":["stocks"],"summary":"Update Stock Price","description":"Manipulate stock price.","operationId":"update_stock_price_stocks__ticker__price_post","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"number","title":"Price"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stocks/{ticker}/snapshots":{"get":{"tags":["stocks"],"summary":"Get Stock Snapshots","description":"Get stock price snapshots for graphing.","operationId":"get_stock_snapshots_stocks__ticker__snapshots_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":30,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StockSnapshotResponse"},"title":"Response Get Stock Snapshots Stocks  Ticker  Snapshots Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/stocks/{ticker}/events":{"get":{"tags":["stocks"],"summary":"Get Stock Events","description":"Get price change events (activity log).","operationId":"get_stock_events_stocks__ticker__events_get","parameters":[{"name":"ticker","in":"path","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PriceEventResponse"},"title":"Response Get Stock Events Stocks  Ticker  Events Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/swipe/":{"post":{"tags":["swipe"],"summary":"Swipe","description":"Record a swipe and update stock value.\n\nUses per-ticker locking to prevent race conditions when multiple users\nswipe the same stock simultaneously.","operationId":"swipe_swipe__post","parameters":[{"name":"ticker","in":"query","required":true,"schema":{"type":"string","title":"Ticker"}},{"name":"direction","in":"query","required":true,"schema":{"$ref":"#/components/schemas/SwipeDirection"}},{"name":"token","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwipeResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/generate/description":{"post":{"tags":["ai"],"summary":"Generate Description","description":"Start generating a stock description.","operationId":"generate_description_ai_generate_description_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateDescriptionRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITaskCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/generate/image":{"post":{"tags":["ai"],"summary":"Generate Image","description":"Start generating an image for a stock.","operationId":"generate_image_ai_generate_image_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateImageRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITaskCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/generate/video":{"post":{"tags":["ai"],"summary":"Generate Video","description":"Start generating a video ad for a stock.","operationId":"generate_video_ai_generate_video_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateVideoRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITaskCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/generate/headlines":{"post":{"tags":["ai"],"summary":"Generate Headlines","description":"Generate satirical news headlines about the top volatile stocks.\n\nReturns headlines immediately (synchronous generation).","operationId":"generate_headlines_ai_generate_headlines_post","parameters":[{"name":"count","in":"query","required":false,"schema":{"type":"integer","default":5,"title":"Count"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HeadlinesResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/generate/stock-groups":{"get":{"tags":["ai"],"summary":"Generate Stock Groups","description":"Generate AI-created sector groupings for random stocks.\n\nReturns groups of stocks with satirical sector names for sunburst visualization.","operationId":"generate_stock_groups_ai_generate_stock_groups_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StockGroupsResponse"}}}}}}},"/ai/tasks":{"get":{"tags":["ai"],"summary":"List Tasks","description":"List all AI tasks, optionally filtered by status or type.","operationId":"list_tasks_ai_tasks_get","parameters":[{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskStatus"},{"type":"null"}],"title":"Status"}},{"name":"task_type","in":"query","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/TaskType"},{"type":"null"}],"title":"Task Type"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AITaskResponse"},"title":"Response List Tasks Ai Tasks Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/tasks/{task_id}":{"get":{"tags":["ai"],"summary":"Get Task","description":"Get a specific AI task by ID.","operationId":"get_task_ai_tasks__task_id__get","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AITaskResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["ai"],"summary":"Delete Task","description":"Delete an AI task.","operationId":"delete_task_ai_tasks__task_id__delete","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ai/tasks/{task_id}/apply":{"post":{"tags":["ai"],"summary":"Apply Result","description":"Apply a completed task's result to a stock.","operationId":"apply_result_ai_tasks__task_id__apply_post","parameters":[{"name":"task_id","in":"path","required":true,"schema":{"type":"string","title":"Task Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApplyResultRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/screenshot/views":{"get":{"tags":["screenshot"],"summary":"List Views","description":"List available screenshot views.","operationId":"list_views_screenshot_views_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"boolean"}]},"type":"object","title":"Response List Views Screenshot Views Get"}}}}}}},"/screenshot/{view}.jpg":{"get":{"tags":["screenshot"],"summary":"Get Screenshot","description":"Get a single screenshot as JPEG.\n\nFast endpoint for Pi to poll - service starts lazily on first request.","operationId":"get_screenshot_screenshot__view__jpg_get","parameters":[{"name":"view","in":"path","required":true,"schema":{"type":"string","description":"View name to capture","title":"View"},"description":"View name to capture"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/screenshot/stream/{view}":{"get":{"tags":["screenshot"],"summary":"Stream View","description":"Stream a view as MJPEG.\n\nPi can open this directly with mpv or ffplay:\n   mpv --no-cache http://localhost:8080/api/screenshot/stream/leaderboard","operationId":"stream_view_screenshot_stream__view__get","parameters":[{"name":"view","in":"path","required":true,"schema":{"type":"string","description":"View name to stream","title":"View"},"description":"View name to stream"},{"name":"fps","in":"query","required":false,"schema":{"type":"number","maximum":30,"minimum":0.1,"description":"Target FPS","default":5.0,"title":"Fps"},"description":"Target FPS"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/screenshot/reload":{"post":{"tags":["screenshot"],"summary":"Reload All Views","description":"Reload all view pages (useful after frontend changes).","operationId":"reload_all_views_screenshot_reload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":{"type":"string"},"type":"object","title":"Response Reload All Views Screenshot Reload Post"}}}}}}},"/screenshot/reload/{view}":{"post":{"tags":["screenshot"],"summary":"Reload View","description":"Reload a specific view page.","operationId":"reload_view_screenshot_reload__view__post","parameters":[{"name":"view","in":"path","required":true,"schema":{"type":"string","description":"View name to reload","title":"View"},"description":"View name to reload"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"title":"Response Reload View Screenshot Reload  View  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Health check endpoint with system status.\n\nReturns:\n    - status: \"ok\" if all checks pass, \"degraded\" if some fail\n    - database: connectivity status\n    - scheduler: running status and job count\n    - disk: free space in the images directory","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AITaskCreateResponse":{"properties":{"task_id":{"type":"string","title":"Task Id"},"status":{"$ref":"#/components/schemas/TaskStatus"},"message":{"type":"string","title":"Message"}},"type":"object","required":["task_id","status","message"],"title":"AITaskCreateResponse","description":"Response when creating an AI task."},"AITaskResponse":{"properties":{"id":{"type":"string","title":"Id"},"ticker":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ticker"},"task_type":{"$ref":"#/components/schemas/TaskType"},"status":{"$ref":"#/components/schemas/TaskStatus"},"prompt":{"type":"string","title":"Prompt"},"model":{"type":"string","title":"Model"},"image_type":{"anyOf":[{"$ref":"#/components/schemas/ImageType"},{"type":"null"}]},"result":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Result"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"completed_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Completed At"}},"type":"object","required":["id","ticker","task_type","status","prompt","model","image_type","result","error","created_at","completed_at"],"title":"AITaskResponse","description":"Response schema for an AI task."},"ApplyResultRequest":{"properties":{"ticker":{"type":"string","title":"Ticker"}},"type":"object","required":["ticker"],"title":"ApplyResultRequest","description":"Request to apply a generated result to a stock."},"Body_create_stock_stocks__post":{"properties":{"ticker":{"type":"string","title":"Ticker"},"title":{"type":"string","title":"Title"},"description":{"type":"string","title":"Description","default":""},"initial_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Initial Price"},"image":{"anyOf":[{"type":"string","format":"binary"},{"type":"null"}],"title":"Image"}},"type":"object","required":["ticker","title"],"title":"Body_create_stock_stocks__post"},"Body_upload_stock_image_stocks__ticker__image_post":{"properties":{"image":{"type":"string","format":"binary","title":"Image"}},"type":"object","required":["image"],"title":"Body_upload_stock_image_stocks__ticker__image_post"},"ChangeType":{"type":"string","enum":["initial","swipe_up","swipe_down","random","admin"],"title":"ChangeType","description":"Type of price change."},"GenerateDescriptionRequest":{"properties":{"ticker":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ticker"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"}},"type":"object","title":"GenerateDescriptionRequest","description":"Request to generate/modify a stock description."},"GenerateImageRequest":{"properties":{"ticker":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ticker"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"image_type":{"$ref":"#/components/schemas/ImageType","default":"main"},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"}},"type":"object","title":"GenerateImageRequest","description":"Request to generate an image for a stock."},"GenerateVideoRequest":{"properties":{"ticker":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ticker"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"source_image_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source Image Url"},"duration":{"type":"integer","title":"Duration","default":5},"model":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model"}},"type":"object","title":"GenerateVideoRequest","description":"Request to generate a video ad for a stock."},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"HeadlinesResponse":{"properties":{"headlines":{"items":{"type":"string"},"type":"array","title":"Headlines"},"stocks_used":{"items":{"type":"string"},"type":"array","title":"Stocks Used"}},"type":"object","required":["headlines","stocks_used"],"title":"HeadlinesResponse","description":"Response containing generated headlines."},"ImageType":{"type":"string","enum":["main","logo","billboard","website"],"title":"ImageType","description":"Type of image to generate."},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"},"note":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Note"}},"type":"object","required":["message"],"title":"MessageResponse","description":"Simple message response for operations without complex data."},"PriceEventResponse":{"properties":{"id":{"type":"integer","title":"Id"},"price":{"type":"number","title":"Price"},"change_type":{"$ref":"#/components/schemas/ChangeType"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","price","change_type","created_at"],"title":"PriceEventResponse","description":"Price change event for API response."},"StockGroup":{"properties":{"name":{"type":"string","title":"Name"},"stocks":{"items":{"$ref":"#/components/schemas/StockInGroup"},"type":"array","title":"Stocks"}},"type":"object","required":["name","stocks"],"title":"StockGroup","description":"A sector group with AI-generated name and stocks."},"StockGroupsResponse":{"properties":{"groups":{"items":{"$ref":"#/components/schemas/StockGroup"},"type":"array","title":"Groups"}},"type":"object","required":["groups"],"title":"StockGroupsResponse","description":"Response containing AI-generated stock groupings."},"StockInGroup":{"properties":{"ticker":{"type":"string","title":"Ticker"},"title":{"type":"string","title":"Title"},"price":{"type":"number","title":"Price"},"percent_change":{"type":"number","title":"Percent Change"}},"type":"object","required":["ticker","title","price","percent_change"],"title":"StockInGroup","description":"A stock within a sector group."},"StockOrder":{"type":"string","enum":["default","random","rank","rank_desc","created_at","created_at_desc","change_rank","change_rank_desc"],"title":"StockOrder","description":"Stock ordering options."},"StockResponse":{"properties":{"ticker":{"type":"string","title":"Ticker"},"title":{"type":"string","title":"Title"},"image":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Image"},"description":{"type":"string","title":"Description"},"is_active":{"type":"boolean","title":"Is Active"},"price":{"type":"number","title":"Price"},"max_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Price"},"min_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Price"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"reference_price":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Reference Price"},"reference_price_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Reference Price At"},"percentage_change":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Percentage Change"},"rank":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rank"},"previous_rank":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Previous Rank"},"rank_change":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Rank Change"},"change_rank":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Change Rank"},"previous_change_rank":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Previous Change Rank"},"change_rank_change":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Change Rank Change"},"initial_price":{"type":"number","title":"Initial Price","description":"Get reference price (from last snapshot) as base for change calculation.","readOnly":true},"change":{"type":"number","title":"Change","description":"Absolute change from reference price.","readOnly":true},"percent_change":{"type":"number","title":"Percent Change","description":"Percentage change from reference price.","readOnly":true}},"type":"object","required":["ticker","title","image","description","is_active","price","created_at","updated_at","initial_price","change","percent_change"],"title":"StockResponse","description":"Stock response matching frontend interface."},"StockSnapshotResponse":{"properties":{"price":{"type":"number","title":"Price"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["price","created_at"],"title":"StockSnapshotResponse","description":"Price snapshot for graphs."},"SwipeDirection":{"type":"string","enum":["left","right"],"title":"SwipeDirection","description":"Swipe direction."},"SwipeResponse":{"properties":{"ticker":{"type":"string","title":"Ticker"},"new_price":{"type":"number","title":"New Price"},"delta":{"type":"number","title":"Delta"},"token":{"type":"string","title":"Token"}},"type":"object","required":["ticker","new_price","delta","token"],"title":"SwipeResponse","description":"Response after a swipe action."},"TaskStatus":{"type":"string","enum":["pending","processing","completed","failed"],"title":"TaskStatus","description":"Status of an AI task."},"TaskType":{"type":"string","enum":["description","image","video"],"title":"TaskType","description":"Type of AI generation task."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}